requirements.txt
# Core
pandas
numpy
sqlalchemy
psycopg2-binary
python-dotenv
pyyaml

# AI / LLM
openai

# Streamlit App
streamlit
plotly

# Analysis
matplotlib
tabulate

# Testing
pytest
